Implementing a switch based sectionControl - based on the code by MTZ8302
I was inspired by the solution of MTZ8302 to build a section-control-box for my equipment. I found his solution very interesting, but the switching (MAN/AUTO,... ) logic is not exactly fitting my needs at the moment. I prefer to have section controlling priority with switches – not with buttons - so I can feel the selected state even without looking at the indication (switch-box or AOG). To toggle the AGO section state (ON/AUTO/OFF) on the screen is not necessary in my case.

Each Section has an ON / AUTO / OFF Position (3 position switch):
The (top) ON position is always setting the section to ON, as long as the MAIN switch in not avoiding that. So it is manually overriding the AOG output to ON.

The (middle) AUTO position is setting the section to the incoming AOG request, as long as the MAIN switch in ON.

The OFF position is setting the section always to OFF. So it is manually overriding the AOG output to OFF.

I am additional using on MAIN Switch with an ON / OFF / EXTERNAL position (3 position switch – 2 positions is possible if the EXTERNAL option is not needed):
The ON position is giving the switch-box the possibility to select all sections to ON, if they are in (top) ON or in the (middle) AUTO position, when an ON signal is received from AOG.

The OFF position is switching all sections to OFF regardless of anything else.

The EXTERNAL position (bottom) has more or less the ON functionality, but with the additional necessarily to have an “external” switch in the ON position as well, to giving the switch-box the possibility to switch all sections ON.
This external switch could be used as an overriding safety-master-switch mounted on very convenient ( quick to reach) location, e.g. if you are not happy with the auto mode during tight turns in problematic corners of the field or during “offline” usage of the switch-box to avoid the need of reaching out to the switch-box behind you, … .
It could well be used like the work switch of the equipment, to avoid “painting” when the tool is not in the right configuration or working position.

So my solution is this PR:
Additional to the initial buttons solution by MTZ8302 - ( both hardware setups are possible )- the “new” hardware is sending a bit in its standard PGN234 to select the way AGO will be processing the received information. The old "unused" 3rd bit of the main byte (5th byte) is now used for this. MTZ8302 existing hardware is/was sending 0, in my solution a 1 is sent.

Byte 6 and 7 (old unused “Reserved” bytes) of PGN234 are now used to transmit the "Auto"-state of the sections switches. The transmitting style is the same way as the existing On and OFF Bytes are done.
Now, for the section/switch control hardware, it is possible to inform AGO about the exactly desired state of the sections/switches!
The section-control-box keeps controlling the tool in “offline” mode (no AOG connection) – during connection problems or even on a different tractor without AOG tablet at all.